Clean up the display
Implemented a log file where most of the diagnostic output can be
placed. This let me clean up the console output to be more user friendly

There is still mork work to do like making it more clear when we are
Waiting for all of the test results to download.